ras: fix crash caused by unsigned Bpp (class md_vdp)
A recent change made Bpp unsigned. Unfortunately, Bpp must be signed for
computations in ras.cpp to work properly. This patch adds the necessary
type casting.

+++ ras-drawplane.h
@@ -67,7 +67,7 @@
   tile_line = tiles + xsize * yoff;
   scan = yscroll & 7;
 
-  where = dest + (xstart + (xscroll & 7)) * Bpp;
+  where = dest + (xstart + (xscroll & 7)) * (int)Bpp;
 
   for(x = -1; x < w; ++x)
     {
